native FFT's need some tweaks when/if dealing with long double in order to approach FFTW3 accuracy, definition of pi matters, but M_PIl maybe a non-standard GNU extension

+++ source/cpp_src/Operators/test/TestMultidimensionalFastFourierTransform.cc
@@ -12,11 +12,10 @@
 using namespace hops;
 
 #ifdef HOPS_USE_FFTW3
-//#define FFT_TYPE MHO_MultidimensionalFastFourierTransformFFTW<3>
-typedef float FPTYPE;
+typedef long double FPTYPE;
 #define FFT_TYPE MHO_MultidimensionalFastFourierTransformFFTW<FPTYPE,3>
 #else
-typedef float FPTYPE;
+typedef long double FPTYPE;
 #define FFT_TYPE MHO_MultidimensionalFastFourierTransform<FPTYPE,3>
 #endif
